+++ packages/core/src/components/scrollable/scrollable.interfaces.ts
@@ -1,0 +1,5 @@
+export type ScrollPosition = "noScroll" | "top" | "bottom" | "middle";
+
+export interface DsoScrollEndEvent {
+  scrollEnd: "top" | "bottom";
+}