+++ packages/dso-toolkit/src/components/scrollable/scrollable.models.ts
@@ -1,0 +1,8 @@
+export interface Scrollable<TemplateFnReturnType> {
+  dsoScrollEnd?: (e: CustomEvent<DsoScrollEndEvent>) => void;
+  children: TemplateFnReturnType;
+}
+
+export interface DsoScrollEndEvent {
+  scrollEnd: "top" | "bottom";
+}