+++ packages/react/src/components/scrollable/scrollable.react-template.tsx
@@ -1,0 +1,14 @@
+import { Scrollable } from "dso-toolkit";
+import * as React from "react";
+
+import { DsoScrollable } from "../../components";
+import { ComponentImplementation } from "../../templates";
+
+export const reactScrollable: ComponentImplementation<Scrollable<JSX.Element>> = {
+  component: "scrollable",
+  implementation: "react",
+  template: () =>
+    function scrollableTemplate({ children, dsoScrollEnd }) {
+      return <DsoScrollable onDsoScrollEnd={dsoScrollEnd}>{children}</DsoScrollable>;
+    },
+};