+++ packages/react/src/components/scrollable/scrollable.stories.ts
@@ -1,0 +1,26 @@
+import { storiesOfScrollable } from "dso-toolkit";
+import { storiesOf } from "@storybook/react";
+import { templateContainer } from "../../templates";
+
+import readme from "./readme.md";
+import { decorator } from "./scrollable.decorator";
+import { defaultContent, dynamicContent } from "./scrollable.content";
+
+storiesOfScrollable(
+  {
+    parameters: {
+      module,
+      storiesOf,
+      readme,
+    },
+    templateContainer,
+    storyTemplates: ({ scrollableTemplate }, templates) => ({
+      scrollableTemplate,
+      defaultContent,
+      dynamicContent: dynamicContent(templates),
+    }),
+  },
+  {
+    decorator,
+  }
+);