+++ storybook/src/templates.ts
@@ -71,6 +71,7 @@ import {
   ResponsiveElement,
   RichContent,
   RowEqualHeights,
+  Scrollable,
   SearchBar,
   Selectable,
   ShoppingCart,
@@ -190,6 +191,7 @@ import { cssRowEqualHeights } from "./components/row-equal-heights/row-equal-hei
 import { coreSlideToggle } from "./components/slide-toggle/slide-toggle.core-template";
 import { cssRichContent } from "./components/rich-content/rich-content.css-template";
 import { coreActionList } from "./components/action-list/action-list.core-template";
+import { coreScrollable } from "./components/scrollable/scrollable.core-template";
 
 export interface Components {
   accordion: Accordion<TemplateResult>;
@@ -259,8 +261,9 @@ export interface Components {
   progressBar: ProgressBar;
   progressIndicator: ProgressIndicator;
   responsiveElement: ResponsiveElement<TemplateResult>;
-  rowEqualHeights: RowEqualHeights<TemplateResult>;
   richContent: RichContent<TemplateResult>;
+  rowEqualHeights: RowEqualHeights<TemplateResult>;
+  scrollable: Scrollable<TemplateResult>;
   searchBar: SearchBar;
   selectable: Selectable<TemplateResult>;
   shoppingCart: ShoppingCart;
@@ -335,6 +338,7 @@ templateContainer.add(corePagination);
 templateContainer.add(coreProgressBar);
 templateContainer.add(coreProgressIndicator);
 templateContainer.add(coreResponsiveElement);
+templateContainer.add(coreScrollable);
 templateContainer.add(coreSelectable);
 templateContainer.add(coreSlideToggle);
 templateContainer.add(coreTable);